+++ packages/query-core/src/query.ts
@@ -260,9 +260,13 @@ export class Query<
     this.cancel({ silent: true })
   }
 
+  get resetState(): QueryState<TData, TError> {
+    return this.#initialState
+  }
+
   reset(): void {
     this.destroy()
-    this.setState(this.#initialState)
+    this.setState(this.resetState)
   }
 
   isActive(): boolean {
@@ -276,10 +280,11 @@ export class Query<
       return !this.isActive()
     }
     // if a query has no observers, it should still be considered disabled if it never attempted a fetch
-    return (
-      this.options.queryFn === skipToken ||
-      this.state.dataUpdateCount + this.state.errorUpdateCount === 0
-    )
+    return this.options.queryFn === skipToken || !this.isFetched()
+  }
+
+  isFetched() {
+    return this.state.dataUpdateCount + this.state.errorUpdateCount > 0
   }
 
   isStatic(): boolean {
@@ -359,7 +364,7 @@ export class Query<
         // If the transport layer does not support cancellation
         // we'll let the query continue so the result can be cached
         if (this.#retryer) {
-          if (this.#abortSignalConsumed) {
+          if (this.#abortSignalConsumed || this.#isInitialPausedFetch()) {
             this.#retryer.cancel({ revert: true })
           } else {
             this.#retryer.cancelRetry()
@@ -375,6 +380,12 @@ export class Query<
 
   getObserversCount(): number {
     return this.observers.length
+  }
+
+  #isInitialPausedFetch(): boolean {
+    return (
+      this.state.fetchStatus === 'paused' && this.state.status === 'pending'
+    )
   }
 
   invalidate(): void {