+++ packages/query-core/src/utils.ts
@@ -86,6 +86,9 @@ export type QueryTypeFilter = 'all' | 'active' | 'inactive'
 
 // UTILS
 
+/** @deprecated
+ * use `environmentManager.isServer()` instead.
+ */
 export const isServer = typeof window === 'undefined' || 'Deno' in globalThis
 
 export function noop(): void