+++ packages/react-query-next-experimental/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @tanstack/react-query-next-experimental
 
+## 5.95.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @tanstack/react-query@5.95.0
+
+## 5.94.5
+
+### Patch Changes
+
+- fix(\*): resolve issue about excluded build directory ([#10312](https://github.com/TanStack/query/pull/10312))
+
+- Updated dependencies [[`4b6536d`](https://github.com/TanStack/query/commit/4b6536dfce99036f4e37f52943c6fed3ad0e0a18)]:
+  - @tanstack/react-query@5.94.5
+
+## 5.94.4
+
+### Patch Changes
+
+- chore: fixed version ([#10064](https://github.com/TanStack/query/pull/10064))
+
+- Updated dependencies [[`4c75210`](https://github.com/TanStack/query/commit/4c75210ce8235fe3d39b67e1029eff11278927cc)]:
+  - @tanstack/react-query@5.94.4
+
 ## 5.91.0
 
 ### Minor Changes